backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI 文件助理的後端伺服器，負責處理文件、AI 分析和 API 服務。"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "motor>=3.0.0",
    "python-dotenv>=1.0.0",
    # "httpx>=0.28.1", # 主要用於測試，已在 test optional-dependencies 中
    "pydantic-settings>=2.9.1",
    "pytest-asyncio>=0.26.0",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # 向量資料庫和Embedding相關依賴 - 使用 ChromaDB (對 Windows 支援更好)
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.5",
    "httpx>=0.23.0", # 確保 httpx 在測試依賴中 (0.28.1 也可以)
]

[tool.pytest.ini_options]
# asyncio_mode = "auto" # 移除此行，因為 pytest-asyncio 通常會自動處理或默認為 auto
# 如果需要，可以添加其他 pytest 選項，例如:
addopts = "-ra -q --show-capture=no" # -ra: 顯示詳細的失敗信息, -q: 安靜模式, --show-capture=no: 不捕獲標準輸出
testpaths = [
    "tests",
]
pythonpath = [
    ".",
    ".."
]
markers = [
    "asyncio: 標記異步測試 (由 pytest-asyncio 提供)",
]
